src/main/webapp/app/entities/bill-table/delete/bill-table-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { BillTableService } from '../service/bill-table.service';

import { BillTableDeleteDialogComponent } from './bill-table-delete-dialog.component';

describe('BillTable Management Delete Component', () => {
  let comp: BillTableDeleteDialogComponent;
  let fixture: ComponentFixture<BillTableDeleteDialogComponent>;
  let service: BillTableService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, BillTableDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(BillTableDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(BillTableDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(BillTableService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
